feat: refactor imports for FAQPage, ContactPage, and HomePage; consolidate AboutUs and Hero components
