+++ frontend/src/app/hero-1/index.tsx
