+++ frontend/src/pages/HomePage.tsx
@@ -1,6 +1,6 @@
-import Features02Page from "@/app/AboutUs/AboutUs"
-import Footer05Page from "@/app/Footer/Footer"
-import Hero01 from "@/app/hero-1/Page"
+import Features02Page from "@/app/AboutUs"
+import Footer05Page from "@/app/Footer"
+import Hero01 from "@/app/hero-1"
 import Navbar04Page from "@/app/navbar/page"
 
 const HomePage = () => {
